fix(reports): scan payment aggregate columns as float64, not int64
GetAgingBuckets and GetMonthlyCollectionTrend scanned numeric(10,2)
SUM() results into int64 Go variables. lib/pq returns numeric columns
as []uint8 byte strings ("1000.00"), which fails to parse as an int64
on the decimal point, and SUM() over zero matching rows returns NULL
with no matching rows to sum, which also fails to scan into a
non-nullable type. Changed the aggregate fields through the full
chain (model, interface, repository query with COALESCE, mock) to
float64.

+++ src/backend/api/internal/handlers/report_handler_test.go
@@ -115,7 +115,7 @@ func sampleCollectionSummaryReport() *models.CollectionSummaryReport {
 		TotalCollected: 151000,
 		TotalPending:   49000,
 		TotalOverdue:   10000,
-		AgingBuckets:   map[string]int64{"0-30": 5000, "31-60": 3000, "60+": 2000},
+		AgingBuckets:   map[string]float64{"0-30": 5000, "31-60": 3000, "60+": 2000},
 		MonthlyTrend:   []*models.MonthlyCollectionTrend{},
 		ReportPeriod:   "2026-05-01 to 2026-06-01",
 		GeneratedAt:    time.Now(),
